LayoutTests/imported/w3c/web-platform-tests/webcodecs/full-cycle-test.https.any_h264_avc-expected.txt

CONSOLE MESSAGE: DataError: Key frame is required
CONSOLE MESSAGE: DataError: Key frame is required

Harness Error (FAIL), message = DataError: Key frame is required

TIMEOUT Encoding and decoding cycle Test timed out

Harness Error (FAIL), message = DataError: Key frame is required

TIMEOUT Encoding and decoding cycle Test timed out
